feat(footer.html, style.css, index.html, main.js): add Rick Astley easter egg
Add a clickable Rick Astley icon to the footer that plays "Never Gonna Give You Up" when clicked. Add a new class to the footer to contain the social links and the new Rick Astley icon. Add a new class to the footer to contain the actual footer content. Add a new CSS animation to the Rick Astley icon. Add a new link to the footer that redirects to the YouTube video. Add a new script tag to the index.html file to load the YouTube Iframe API. Add a new script to the main.js file to create a new YouTube player and embed it in the page when the Rick Astley link is clicked.